Extracts StandardItem into a separate file
